+++ src/Command/MatiStreamCommand.php
@@ -5,15 +5,15 @@ declare(strict_types=1);
 namespace Mati\Command;
 
 use Doctrine\ORM\EntityManagerInterface;
-use Mati\Dto\IpcMessage;
+use Mati\Ipc\IpcMessage;
 use Mati\Ipc\IpcServer;
 use Mati\Ipc\Terminator;
+use Mati\Livestream\ChatClient;
+use Mati\Livestream\ChatUrlFetcher;
+use Mati\Livestream\LivestreamUrlFetcher;
 use Mati\MatiConfiguration;
 use Mati\Repository\StreamRepository;
 use Mati\Repository\SuperchatRepository;
-use Mati\Rumble\ChatClient;
-use Mati\Rumble\ChatUrlFetcher;
-use Mati\Rumble\LivestreamUrlFetcher;
 use Mati\Superchat\SuperchatCache;
 use Mati\Superchat\SuperchatConverter;
 use Psr\Log\LoggerInterface;
